requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
transformers==4.38.2
torch==2.2.0
faiss-cpu==1.7.4
langchain==0.1.9
python-dotenv==1.0.0
sentence-transformers==2.5.1
requests==2.31.0
groq==0.4.2